Refactor: lift lambda tyarg Class obj construction to caller
Move the construction of runtime Class objects for lambda type
arguments out of build_fn_object and into the HirLambdaExpr handler.
build_fn_object now receives the pre-built Vec<TypedExpr>.

This is a pure refactor; TyPara cases remain todo!() and behavior
is unchanged.
